Cleaing project directory by renewing folder names and file names, while also removing unnecessary dummy files. Also added an image file which shows the project directory structure in my VS Code workspace.
If you want to run this code, make sure to have the necessary permissions and backup your data before executing it.

Also, the image files are not included in the project directory as the data size is too large to be uploaded here.

The basic assumption is that, all members of the project directory are in the same folder level, and the script will not traverse into subdirectories.

I have uploaded the processed image files in the Google Drive shared folder, which you can access using the link provided in the project documentation (Google Docs).